{"schema":"decodex/commit/1","summary":"Modularize agent evidence","authority":"XY-1115"}
